Moving the translatable strings into typescript structures
